# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Brasseries - CAMEROUN",
        "description": " OpenApi description for Brasseries - CAMEROUN",
        "contact": {
            "name": "CAMTRACK - DEV",
            "url": "https://camtrack.net/",
            "email": "alfred.andrianjatovo@camtrack.mg"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Brasseries API Sources"
        }
    ],
    "paths": {
...
        "/departements": {
            "get": {
                "tags": [
                    "CRUD Departement"
                ],
                "summary": "Get list of departement",
                "description": "Get list of departement",
                "operationId": "indexDepartement",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRUD Departement"
                ],
                "summary": "Store Departement in DB",
                "description": "Store departement in DB",
                "operationId": "storeDepartement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "site_id"
                                ],
                                "properties": {
                                    "site_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/departements/{ids}": {
            "get": {
                "tags": [
                    "CRUD Departement"
                ],
                "summary": "Get Departement Detail",
                "description": "Get departement Detail",
                "operationId": "showDepartement",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of departement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CRUD Departement"
                ],
                "summary": "Update departement in DB",
                "description": "Update departement in DB",
                "operationId": "updateDepartement",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of departement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "site_id"
                                ],
                                "properties": {
                                    "site_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD Departement"
                ],
                "summary": "Delete Departement",
                "description": "Delete departement",
                "operationId": "destroyDepartement",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of departement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/{ids}": {
            "get": {
                "tags": [
                    "CRUD User"
                ],
                "summary": "Get User Detail",
                "description": "Get user Detail",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CRUD User"
                ],
                "summary": "Update user in DB",
                "description": "Update user in DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": ""
                                    },
                                    "cle_user": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "departement_id": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "usertype_id": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD User"
                ],
                "summary": "Delete User",
                "description": "Delete user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication Bearer Token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}